+++ api/v1/trace/data.go
@@ -1,6 +1,8 @@
 package trace
 
-import "strconv"
+import (
+	"strconv"
+)
 
 func string2Float(num string) float64 {
 	r, err := strconv.ParseFloat(num, 64)