+++ api/v1/trace/jma.go
@@ -6,6 +6,7 @@ import (
 	"time"
 
 	"github.com/anyshake/observer/utils/request"
+	"github.com/corpix/uarand"
 )
 
 type JMA struct {
@@ -24,6 +25,7 @@ func (j *JMA) Fetch() ([]byte, error) {
 	res, err := request.GET(
 		"https://www.jma.go.jp/bosai/quake/data/list.json",
 		10*time.Second, time.Second, 3, false, nil,
+		map[string]string{"User-Agent": uarand.GetRandom()},
 	)
 	if err != nil {
 		return nil, err